added calculation of elo

+++ megamek/src/megamek/common/IPlayerWithRating.java
@@ -1,7 +1,7 @@
 package megamek.common;
 
 public interface IPlayerWithRating {
-    public int getEloRating();
-    public int modifyEloRating(int delta);
+    public float getEloRating();
+    public float modifyEloRating(float delta);
     public void resetEloRating();
 }
